+++ NEWS.md
@@ -1,3 +1,5 @@
+# devrcap 0.2.0
+
 * Added `geom_timeline_label()` function to output the ggplot layer for
   the annotation to the earthquakes time line
 * Added `geom_timeline()` function to output the ggplot2 layer for